move LogStatus to pkg/logs/internal/status
Note that this leaves pkg/logs/status alone.  That is a different kind
of status, is global, and is tied in with some host-metadata stuff
outside of pkg/logs, so it can't move to pkg/logs/internal.  See
AML-190.
